ai_space_news_summarizer_structured/requirements.txt
streamlit>=1.31
requests>=2.31
feedparser>=6.0
beautifulsoup4>=4.12
python-dateutil>=2.8
pyyaml>=6.0